Move feature detection to env.js. Move unit test with them.
